requirements.txt
astroid==2.9.0
autopep8==1.6.0
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.9
google-api-core==2.3.0
google-api-python-client==2.33.0
google-auth==2.3.3
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
googleapis-common-protos==1.54.0
httplib2==0.20.2
idna==3.3
isort==5.10.1
lazy-object-proxy==1.6.0
mccabe==0.6.1
oauthlib==3.1.1
platformdirs==2.4.0
protobuf==3.19.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pylint==2.12.2
pyparsing==3.0.6
python-dotenv==0.19.2
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.8
six==1.16.0
spotipy==2.19.0
toml==0.10.2
typing_extensions==4.0.1
uritemplate==4.1.1
urllib3==1.26.7
wrapt==1.13.3
